Removed 'most keys bring down console when playing demo'.  It was a problem for using bound keys during demo playback.  Just hit ESC for console like usual.
Added a silly fisheye zoom command, r_fishaccel.  The amount of zoom will accelerate according to this value.

Some progs.dat (like Marcher Fortress) send static sounds as bytes.  This was causing the text at the elevator near then end of Marcher to crash with super8 protocol, which sent sound as short.  It's a special multitrigger entity.  So static sounds was changed to be compatible.  Other sound can still be sent as short.